+++ app/policies/dashboard_policy.rb
@@ -1,5 +1,0 @@
-class DashboardPolicy < ApplicationPolicy
-  def show?
-    true
-  end
-end