+++ app/policies/home_policy.rb
@@ -1,0 +1,6 @@
+class HomePolicy < ApplicationPolicy
+  def show?
+    return true if beis_user? || delivery_partner_user?
+    false
+  end
+end