+++ spec/features/staff/users_can_view_the_site_navigation_spec.rb
@@ -18,7 +18,7 @@ RSpec.feature "Users can view the site navigation" do
       visit organisation_path(user.organisation)
 
       expect(page).to have_css ".govuk-header__navigation"
-      expect(page).to have_link t("page_title.home"), href: organisation_path(user.organisation)
+      expect(page).to have_link t("page_title.home"), href: home_path
       expect(page).to have_link t("page_title.report.index"), href: reports_path
       expect(page).not_to have_link t("page_title.organisation.index"), href: organisations_path
       expect(page).not_to have_link t("page_title.users.index"), href: users_path
@@ -33,7 +33,7 @@ RSpec.feature "Users can view the site navigation" do
       visit organisation_path(user.organisation)
 
       expect(page).to have_css ".govuk-header__navigation"
-      expect(page).to have_link t("page_title.home"), href: organisation_path(user.organisation)
+      expect(page).to have_link t("page_title.home"), href: home_path
       expect(page).to have_link t("page_title.report.index"), href: reports_path
       expect(page).to have_link t("page_title.organisation.index"), href: organisations_path
       expect(page).to have_link t("page_title.users.index"), href: users_path